requirements.txt
requests
google-generativeai
python-telegram-bot